requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Semantic segmentation models with pre-trained encoders
segmentation-models-pytorch>=0.3.3

# Geospatial libraries
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0
Fiona>=1.9.0
mercantile>=1.2.1

# Image processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data handling & numerics
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration & utilities
PyYAML>=6.0
tqdm>=4.65.0
rich>=13.0.0
beautifulsoup4>=4.12.0

# ---------------------------------------------------------------------------
# Satellite data acquisition
# ---------------------------------------------------------------------------

# Microsoft Planetary Computer (free, no account needed)
planetary-computer>=1.0.0
pystac-client>=0.7.0
stackstac>=0.4.3

# Google Earth Engine (free account required)
earthengine-api>=0.1.370

# ---------------------------------------------------------------------------
# FastAPI server + visualisation
# ---------------------------------------------------------------------------

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9       # required for file upload endpoints
Jinja2>=3.1.0                 # HTML template rendering
aiofiles>=23.2.1

# XYZ tile server (COG-aware, fast)
rio-tiler>=6.0.0

# Optional: multi-worker production server
# gunicorn>=21.2.0

# ---------------------------------------------------------------------------
# Jupyter (notebooks)
# ---------------------------------------------------------------------------
jupyter>=1.0.0
ipykernel>=6.0.0

# ---------------------------------------------------------------------------
# Testing
# ---------------------------------------------------------------------------
pytest>=7.4.0
httpx>=0.27.0                 # required by FastAPI TestClient
